# docs/lifecycle/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "robotsix-central-deploy Lifecycle API",
    "description": "Start, stop, restart, and inspect suite services.",
    "version": "0.1.0",
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/chat/github/repos/{owner}/{repo}/vulnerability-alerts": {
      "put": {
        "tags": [
          "chat-github"
        ],
        "summary": "Enable the Dependency graph / Dependabot vulnerability alerts",
        "description": "Enable *owner*/*repo*'s Dependency graph and Dependabot alerts.\n\nWithout this, GitHub Actions checks that rely on the dependency graph\n(e.g. ``dependency-review-action``) fail with \"Dependency review is not\nsupported on this repository\" until a human visits Settings > Security.\nNew repos get this automatically via :func:`create_repo`; this endpoint\ncovers repos that predate that, or ones created outside the chat agent.",
        "operationId": "enable_vulnerability_alerts_chat_github_repos__owner___repo__vulnerability_alerts_put",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Enable Vulnerability Alerts Chat Github Repos  Owner   Repo  Vulnerability Alerts Put"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Repository not found or App not installed on it"
          },
          "503": {
            "description": "GitHub App not configured"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat/github/repos/{owner}/{repo}/security-features": {
      "put": {
        "tags": [
          "chat-github"
        ],
        "summary": "Toggle repository security features",
        "description": "Enable or disable security features on *owner*/*repo*.\n\nThis combines what would otherwise require two separate GitHub API\ncalls (``vulnerability-alerts`` and ``automated-security-fixes``)\ninto one endpoint.  Only the features whose body fields are provided\nare changed; omit a field to leave it as-is.",
        "operationId": "set_security_features_chat_github_repos__owner___repo__security_features_put",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecurityFeaturesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Set Security Features Chat Github Repos  Owner   Repo  Security Features Put"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Repository not found or App not installed on it"
          },
          "422": {
            "description": "No fields provided, or GitHub rejected the request"
          },
          "503": {
            "description": "Neither GitHub App nor PAT configured for this operation"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SecurityFeaturesRequest": {
        "properties": {
          "dependency_graph": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependency Graph",
            "description": "Enable/disable the Dependency graph (shares the vulnerability-alerts endpoint with Dependabot alerts)."
          },
          "dependabot_alerts": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependabot Alerts",
            "description": "Enable/disable Dependabot vulnerability alerts (shares the vulnerability-alerts endpoint with the Dependency graph)."
          },
          "dependabot_security_updates": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependabot Security Updates",
            "description": "Enable/disable Dependabot automated security fixes."
          }
        },
        "type": "object",
        "title": "SecurityFeaturesRequest",
        "description": "Body for ``PUT /chat/github/repos/{owner}/{repo}/security-features``.\n\nAll three fields are optional; only the ones provided are changed.\n*dependency_graph* and *dependabot_alerts* share the same underlying\nGitHub API (``vulnerability-alerts``), so setting either to ``true``\nenables vulnerability alerts and setting either to ``false`` (without\nthe other being ``true``) disables them.  *dependabot_security_updates*\ncontrols automated security fixes independently."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
  "security": [],
  "servers": [
    {
      "url": "http://localhost:8100",
      "description": "Local development server"
    }
  ]
}